Removing unnecessary loops and changed ints to bytes

+++ src/PuzzleFileReader.java
@@ -5,22 +5,22 @@ import java.util.Scanner;
 
 public class PuzzleFileReader {
 
-    public int[][] readFile(String filename) throws IOException {
+    public byte[][] readFile(String filename) throws IOException {
         Scanner sc = new Scanner(new BufferedReader(new FileReader(filename)));
-        int[][] result = null;
-        int i = 0, length = 0;
+        byte[][] result = null;
+        byte i = 0, length = 0;
         String[] line;
         while (sc.hasNextLine()) {
             line = sc.nextLine().split(" ");
 
             // Assuming that board is always a square
             if (result == null) {
-                length = line.length;
-                result = new int[length][length];
+                length = (byte)line.length;
+                result = new byte[length][length];
             }
 
             for (int j = 0; j < length; j++) {
-                result[i][j] = Integer.parseInt(line[j]);
+                result[i][j] = Byte.parseByte(line[j]);
             }
             ++i;
         }
